LPD-78240 portal-search-text-embeddings: Move TextEmbeddingProvider implementations to dxp module
https://liferay.atlassian.net/browse/LPD-78240
