+++ modules/dxp/apps/portal-search-text-embeddings/portal-search-text-embeddings/src/main/java/com/liferay/portal/search/text/embeddings/HuggingFaceInferenceEndpointTextEmbeddingProvider.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.ml.embedding.text;
+package com.liferay.portal.search.text.embeddings;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -15,10 +16,9 @@ import com.liferay.portal.kernel.servlet.HttpHeaders;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.search.internal.ml.embedding.text.util.ConfigurationValidationUtil;
+import com.liferay.portal.search.ml.embedding.util.ConfigurationValidationUtil;
 import com.liferay.portal.search.rest.dto.v1_0.EmbeddingProviderConfiguration;
-
-import java.net.HttpURLConnection;
+import com.liferay.portal.search.rest.text.embeddings.configuration.TextEmbeddingProvider;
 
 import java.util.List;
 import java.util.Map;
@@ -29,11 +29,8 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Petteri Karttunen
  */
-@Component(
-	property = "provider.name=HuggingFaceInferenceAPI",
-	service = TextEmbeddingProvider.class
-)
-public class HuggingFaceInferenceAPITextEmbeddingProvider
+@Component(service = TextEmbeddingProvider.class)
+public class HuggingFaceInferenceEndpointTextEmbeddingProvider
 	implements TextEmbeddingProvider {
 
 	@Override
@@ -45,7 +42,7 @@ public class HuggingFaceInferenceAPITextEmbeddingProvider
 			(Map<String, Object>)embeddingProviderConfiguration.getAttributes();
 
 		if (!ConfigurationValidationUtil.validateAttributes(
-				attributes, new String[] {"accessToken", "model"})) {
+				attributes, new String[] {"accessToken", "hostAddress"})) {
 
 			return new Double[0];
 		}
@@ -53,25 +50,17 @@ public class HuggingFaceInferenceAPITextEmbeddingProvider
 		return _getEmbedding(attributes, text);
 	}
 
+	@Override
+	public String getProviderName() {
+		return _PROVIDE_NAME;
+	}
+
 	private Double[] _getEmbedding(
 		Map<String, Object> attributes, String text) {
 
 		try {
-			Http.Options options = _getOptions(attributes, text);
-
-			String responseJSON = _http.URLtoString(options);
-
-			Http.Response response = options.getResponse();
-
-			if (response.getResponseCode() ==
-					HttpURLConnection.HTTP_UNAVAILABLE) {
-
-				options.addHeader("x-wait-for-model", "true");
-				options.setTimeout(
-					MapUtil.getInteger(attributes, "modelTimeout", 30) * 1000);
-
-				responseJSON = _http.URLtoString(options);
-			}
+			String responseJSON = _http.URLtoString(
+				_getOptions(attributes, text));
 
 			if (_log.isDebugEnabled()) {
 				_log.debug("Response: " + responseJSON);
@@ -81,9 +70,15 @@ public class HuggingFaceInferenceAPITextEmbeddingProvider
 				throw new IllegalArgumentException(responseJSON);
 			}
 
-			JSONArray jsonArray = _jsonFactory.createJSONArray(responseJSON);
+			JSONArray jsonArray1 = _jsonFactory.createJSONArray(responseJSON);
 
-			List<Double> list = JSONUtil.toDoubleList(jsonArray);
+			JSONArray jsonArray2 = jsonArray1.getJSONArray(0);
+
+			if (jsonArray2 == null) {
+				throw new IllegalArgumentException(responseJSON);
+			}
+
+			List<Double> list = JSONUtil.toDoubleList(jsonArray2);
 
 			return list.toArray(new Double[0]);
 		}
@@ -97,29 +92,33 @@ public class HuggingFaceInferenceAPITextEmbeddingProvider
 
 		Http.Options options = new Http.Options();
 
-		JSONObject jsonObject = JSONUtil.put("inputs", text);
-
 		options.addHeader(
 			HttpHeaders.AUTHORIZATION,
 			"Bearer " + MapUtil.getString(attributes, "accessToken"));
 		options.addHeader(
 			HttpHeaders.CONTENT_TYPE, ContentTypes.APPLICATION_JSON);
 		options.setBody(
-			jsonObject.toString(), ContentTypes.APPLICATION_JSON,
-			StringPool.UTF8);
+			JSONUtil.put(
+				"inputs", text
+			).toString(),
+			ContentTypes.APPLICATION_JSON, StringPool.UTF8);
 		options.setCookieSpec(Http.CookieSpec.STANDARD);
-		options.setLocation(
-			"https://router.huggingface.co/hf-inference/models/" +
-				MapUtil.getString(attributes, "model"));
+		options.setLocation(MapUtil.getString(attributes, "hostAddress"));
 		options.setPost(true);
 
 		return options;
 	}
 
+	private static final String _PROVIDE_NAME =
+		"hugging-face-inference-endpoint";
+
 	private static final Log _log = LogFactoryUtil.getLog(
-		HuggingFaceInferenceAPITextEmbeddingProvider.class);
+		HuggingFaceInferenceEndpointTextEmbeddingProvider.class);
 
 	@Reference
 	private Http _http;
+
+	@Reference
+	private JSONFactory _jsonFactory;
 
 }
