+++ modules/dxp/apps/portal-search-text-embeddings/portal-search-text-embeddings/src/main/java/com/liferay/portal/search/text/embeddings/OpenAITextEmbeddingProvider.java
@@ -3,12 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.ml.embedding.text;
+package com.liferay.portal.search.text.embeddings;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactory;
+import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -16,8 +17,9 @@ import com.liferay.portal.kernel.servlet.HttpHeaders;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.search.internal.ml.embedding.text.util.ConfigurationValidationUtil;
+import com.liferay.portal.search.ml.embedding.util.ConfigurationValidationUtil;
 import com.liferay.portal.search.rest.dto.v1_0.EmbeddingProviderConfiguration;
+import com.liferay.portal.search.rest.text.embeddings.configuration.TextEmbeddingProvider;
 
 import java.util.List;
 import java.util.Map;
@@ -28,12 +30,8 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Petteri Karttunen
  */
-@Component(
-	property = "provider.name=HuggingFaceInferenceEndpoint",
-	service = TextEmbeddingProvider.class
-)
-public class HuggingFaceInferenceEndpointTextEmbeddingProvider
-	implements TextEmbeddingProvider {
+@Component(service = TextEmbeddingProvider.class)
+public class OpenAITextEmbeddingProvider implements TextEmbeddingProvider {
 
 	@Override
 	public Double[] getEmbedding(
@@ -44,7 +42,7 @@ public class HuggingFaceInferenceEndpointTextEmbeddingProvider
 			(Map<String, Object>)embeddingProviderConfiguration.getAttributes();
 
 		if (!ConfigurationValidationUtil.validateAttributes(
-				attributes, new String[] {"accessToken", "hostAddress"})) {
+				attributes, new String[] {"apiKey", "model"})) {
 
 			return new Double[0];
 		}
@@ -52,32 +50,36 @@ public class HuggingFaceInferenceEndpointTextEmbeddingProvider
 		return _getEmbedding(attributes, text);
 	}
 
+	@Override
+	public String getProviderName() {
+		return _PROVIDE_NAME;
+	}
+
 	private Double[] _getEmbedding(
 		Map<String, Object> attributes, String text) {
 
 		try {
-			String responseJSON = _http.URLtoString(
-				_getOptions(attributes, text));
+			JSONObject responseJSONObject = _jsonFactory.createJSONObject(
+				_http.URLtoString(_getOptions(attributes, text)));
+
+			JSONArray dataJSONArray = responseJSONObject.getJSONArray("data");
+
+			if (JSONUtil.isEmpty(dataJSONArray)) {
+				throw new IllegalArgumentException(
+					responseJSONObject.toString());
+			}
+
+			JSONObject dataJSONObject = dataJSONArray.getJSONObject(0);
+
+			List<Double> embedding = JSONUtil.toDoubleList(
+				dataJSONObject.getJSONArray("embedding"));
 
 			if (_log.isDebugEnabled()) {
-				_log.debug("Response: " + responseJSON);
+				_log.debug(
+					"Usage: " + responseJSONObject.getJSONObject("usage"));
 			}
 
-			if (!JSONUtil.isJSONArray(responseJSON)) {
-				throw new IllegalArgumentException(responseJSON);
-			}
-
-			JSONArray jsonArray1 = _jsonFactory.createJSONArray(responseJSON);
-
-			JSONArray jsonArray2 = jsonArray1.getJSONArray(0);
-
-			if (jsonArray2 == null) {
-				throw new IllegalArgumentException(responseJSON);
-			}
-
-			List<Double> list = JSONUtil.toDoubleList(jsonArray2);
-
-			return list.toArray(new Double[0]);
+			return embedding.toArray(new Double[0]);
 		}
 		catch (Exception exception) {
 			return ReflectionUtil.throwException(exception);
@@ -91,23 +93,45 @@ public class HuggingFaceInferenceEndpointTextEmbeddingProvider
 
 		options.addHeader(
 			HttpHeaders.AUTHORIZATION,
-			"Bearer " + MapUtil.getString(attributes, "accessToken"));
+			"Bearer " + MapUtil.getString(attributes, "apiKey"));
 		options.addHeader(
 			HttpHeaders.CONTENT_TYPE, ContentTypes.APPLICATION_JSON);
 		options.setBody(
-			JSONUtil.put(
-				"inputs", text
-			).toString(),
-			ContentTypes.APPLICATION_JSON, StringPool.UTF8);
+			_getRequestBody(attributes, text), ContentTypes.APPLICATION_JSON,
+			StringPool.UTF8);
 		options.setCookieSpec(Http.CookieSpec.STANDARD);
-		options.setLocation(MapUtil.getString(attributes, "hostAddress"));
+		options.setLocation("https://api.openai.com/v1/embeddings");
 		options.setPost(true);
 
 		return options;
 	}
 
+	private String _getRequestBody(
+		Map<String, Object> attributes, String text) {
+
+		JSONObject requestBodyJSONObject = JSONUtil.put(
+			"input", text
+		).put(
+			"model", MapUtil.getString(attributes, "model")
+		);
+
+		if (attributes.containsKey("dimensions")) {
+			requestBodyJSONObject.put(
+				"dimensions", MapUtil.getInteger(attributes, "dimensions"));
+		}
+
+		if (attributes.containsKey("user")) {
+			requestBodyJSONObject.put(
+				"user", MapUtil.getString(attributes, "user"));
+		}
+
+		return requestBodyJSONObject.toString();
+	}
+
+	private static final String _PROVIDE_NAME = "openai";
+
 	private static final Log _log = LogFactoryUtil.getLog(
-		HuggingFaceInferenceEndpointTextEmbeddingProvider.class);
+		OpenAITextEmbeddingProvider.class);
 
 	@Reference
 	private Http _http;
