+++ modules/dxp/apps/portal-search-text-embeddings/portal-search-text-embeddings/src/main/java/com/liferay/portal/search/text/embeddings/TXTAITextEmbeddingProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.ml.embedding.text;
+package com.liferay.portal.search.text.embeddings;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
@@ -15,8 +15,9 @@ import com.liferay.portal.kernel.util.HttpComponentsUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.URLCodec;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.search.internal.ml.embedding.text.util.ConfigurationValidationUtil;
+import com.liferay.portal.search.ml.embedding.util.ConfigurationValidationUtil;
 import com.liferay.portal.search.rest.dto.v1_0.EmbeddingProviderConfiguration;
+import com.liferay.portal.search.rest.text.embeddings.configuration.TextEmbeddingProvider;
 
 import java.util.List;
 import java.util.Map;
@@ -27,9 +28,7 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Petteri Karttunen
  */
-@Component(
-	property = "provider.name=TxtAIText", service = TextEmbeddingProvider.class
-)
+@Component(service = TextEmbeddingProvider.class)
 public class TXTAITextEmbeddingProvider implements TextEmbeddingProvider {
 
 	@Override
@@ -47,6 +46,11 @@ public class TXTAITextEmbeddingProvider implements TextEmbeddingProvider {
 		}
 
 		return _getEmbedding(attributes, text);
+	}
+
+	@Override
+	public String getProviderName() {
+		return _PROVIDE_NAME;
 	}
 
 	private Double[] _getEmbedding(
@@ -101,6 +105,8 @@ public class TXTAITextEmbeddingProvider implements TextEmbeddingProvider {
 
 		return options;
 	}
+
+	private static final String _PROVIDE_NAME = "txtai";
 
 	@Reference
 	private Http _http;
