+++ modules/dxp/apps/portal-search-text-embeddings/portal-search-text-embeddings/src/main/java/com/liferay/portal/search/text/embeddings/VertexAITextEmbeddingProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.ml.embedding.text;
+package com.liferay.portal.search.text.embeddings;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
@@ -18,8 +18,9 @@ import com.liferay.portal.kernel.servlet.HttpHeaders;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.search.internal.ml.embedding.text.util.ConfigurationValidationUtil;
+import com.liferay.portal.search.ml.embedding.util.ConfigurationValidationUtil;
 import com.liferay.portal.search.rest.dto.v1_0.EmbeddingProviderConfiguration;
+import com.liferay.portal.search.rest.text.embeddings.configuration.TextEmbeddingProvider;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -35,9 +36,7 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Petteri Karttunen
  */
-@Component(
-	property = "provider.name=VertexAI", service = TextEmbeddingProvider.class
-)
+@Component(service = TextEmbeddingProvider.class)
 public class VertexAITextEmbeddingProvider implements TextEmbeddingProvider {
 
 	@Override
@@ -55,6 +54,11 @@ public class VertexAITextEmbeddingProvider implements TextEmbeddingProvider {
 		}
 
 		return _getEmbedding(attributes, text);
+	}
+
+	@Override
+	public String getProviderName() {
+		return _PROVIDE_NAME;
 	}
 
 	private String _getAuthenticationToken() throws Exception {
@@ -181,6 +185,8 @@ public class VertexAITextEmbeddingProvider implements TextEmbeddingProvider {
 					MapUtil.getBoolean(attributes, "autoTruncate", true))
 			));
 	}
+
+	private static final String _PROVIDE_NAME = "vertex-ai";
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		VertexAITextEmbeddingProvider.class);
